feat(ecg_dl): 泄漏审计与干净评估管线全量入库（§97–§99）
- 新增 SplitGuard 患者级取数守卫 + audit_provenance RNG 流复放审计（12 历史脚本判 LEAKED）
- 新增干净测试集评估管线（eval_clean_test / eval_clean_baseline / plot_clean_reeval）与 PTB 诊断分层审计
- 新增干净基线从零重训（train_clean_baseline v1/v2，§99：MIT+INCART 首个无泄漏数字 AUC 0.851）
- 归档 08-21~08-31 实验脚本与评估 JSON（ecgfounder/QAT/gate/事件级 sweep 等）
- 文档对齐：TUNING_HISTORY §97–§99、ROADMAP、FINAL_RESULTS、README
- 固件侧：esp_tflm_bench 一致性基准工程与 ads1292r/respiration 模块入库

+++ pc_tools/ecg_dl/models/external_architectures.py
@@ -46,7 +46,8 @@ def build_lstm_cnn_parallel(input_shape=(250, 1), n_classes=2, seed=42):
     x = layers.Concatenate(name="concat")([c, l])
     x = layers.Dense(64, activation="relu", name="dense_1")(x)
     x = layers.Dropout(0.4, name="dropout")(x)
-    outputs = layers.Dense(n_classes, activation="softmax", name="output")(x)
+    outputs = layers.Dense(n_classes, activation="softmax", name="output",
+                           dtype="float32")(x)
 
     model = Model(inputs, outputs, name="lstm_cnn_parallel")
     return model
@@ -82,7 +83,8 @@ def build_cnn_standard(input_shape=(250, 1), n_classes=2, seed=42):
 
     x = layers.Dense(128, activation="relu", name="fc1")(x)
     x = layers.Dropout(0.4, name="dropout")(x)
-    outputs = layers.Dense(n_classes, activation="softmax", name="output")(x)
+    outputs = layers.Dense(n_classes, activation="softmax", name="output",
+                           dtype="float32")(x)
 
     model = Model(inputs, outputs, name="cnn_standard")
     return model
@@ -127,7 +129,8 @@ def build_resnet1d(input_shape=(250, 1), n_classes=2, seed=42):
     x = layers.GlobalAveragePooling1D(name="gap")(x)
     x = layers.Dense(128, activation="relu", name="fc1")(x)
     x = layers.Dropout(0.4, name="dropout")(x)
-    outputs = layers.Dense(n_classes, activation="softmax", name="output")(x)
+    outputs = layers.Dense(n_classes, activation="softmax", name="output",
+                           dtype="float32")(x)
 
     model = Model(inputs, outputs, name="resnet1d")
     return model
